requirements.txt
asgiref==3.8.1
Django==5.0.3
django-leaflet==0.29.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
